+++ app/helpers/namespaces_helper.rb
@@ -1,0 +1,5 @@
+module NamespacesHelper
+  def can_manage_namespace?(namespace)
+    current_user.admin? || namespace.team.owners.exists?(current_user.id)
+  end
+end