+++ packaging/suse/portusctl/lib/configurator.rb
@@ -143,7 +143,7 @@ EOM
   # Creates the secrets.yml file used by Rails
   def secrets
     destination = "/srv/Portus/config/secrets.yml"
-    return if File.exists?(destination)
+    return if File.exist?(destination)
     TemplateWriter.process(
       "secrets.yml.erb",
       destination,